Created an interface for a semantic analyzer (i.e., anything that converts from English to a logical representation)
